it works! (kind of)
