--- assets.dev
@@ -1,8 +1,0 @@
-axe.png
-death_noise.png
-empty.png
-gblin_exp_2.png
-green_square.png
-pickaxe.png
-ts.png
-title_screen_background.png
