--- src/load_maps.rs
@@ -1,5 +1,0 @@
-pub fn load_maps(
-    
-) {
-
-}
