--- test.ldtk
@@ -1,221 +1,0 @@
-{
-	"__header__": {
-		"fileType": "LDtk Project JSON",
-		"app": "LDtk",
-		"doc": "https://ldtk.io/json",
-		"schema": "https://ldtk.io/files/JSON_SCHEMA.json",
-		"appAuthor": "Sebastien 'deepnight' Benard",
-		"appVersion": "0.9.3",
-		"url": "https://ldtk.io"
-	},
-	"jsonVersion": "0.9.3",
-	"nextUid": 5,
-	"worldLayout": "Free",
-	"worldGridWidth": 256,
-	"worldGridHeight": 256,
-	"defaultPivotX": 0,
-	"defaultPivotY": 0,
-	"defaultGridSize": 16,
-	"defaultLevelWidth": 256,
-	"defaultLevelHeight": 256,
-	"bgColor": "#40465B",
-	"defaultLevelBgColor": "#696A79",
-	"minifyJson": false,
-	"externalLevels": false,
-	"exportTiled": false,
-	"imageExportMode": "None",
-	"pngFilePattern": null,
-	"backupOnSave": false,
-	"backupLimit": 10,
-	"levelNamePattern": "Level_%idx",
-	"flags": [],
-	"defs": { "layers": [
-		{
-			"__type": "Entities",
-			"identifier": "Entities",
-			"type": "Entities",
-			"uid": 3,
-			"gridSize": 16,
-			"displayOpacity": 1,
-			"pxOffsetX": 0,
-			"pxOffsetY": 0,
-			"requiredTags": [],
-			"excludedTags": [],
-			"intGridValues": [{ "value": 1, "identifier": null, "color": "#000000" }],
-			"autoTilesetDefUid": null,
-			"autoRuleGroups": [],
-			"autoSourceLayerDefUid": null,
-			"tilesetDefUid": null,
-			"tilePivotX": 0,
-			"tilePivotY": 0
-		},
-		{
-			"__type": "Tiles",
-			"identifier": "Tiles",
-			"type": "Tiles",
-			"uid": 1,
-			"gridSize": 64,
-			"displayOpacity": 1,
-			"pxOffsetX": 0,
-			"pxOffsetY": 0,
-			"requiredTags": [],
-			"excludedTags": [],
-			"intGridValues": [{ "value": 1, "identifier": null, "color": "#000000" }],
-			"autoTilesetDefUid": null,
-			"autoRuleGroups": [],
-			"autoSourceLayerDefUid": null,
-			"tilesetDefUid": 2,
-			"tilePivotX": 0,
-			"tilePivotY": 0
-		}
-	], "entities": [
-		{
-			"identifier": "Player",
-			"uid": 4,
-			"tags": [],
-			"width": 64,
-			"height": 32,
-			"resizableX": false,
-			"resizableY": false,
-			"keepAspectRatio": false,
-			"fillOpacity": 1,
-			"lineOpacity": 1,
-			"hollow": false,
-			"color": "#94D9B3",
-			"renderMode": "Rectangle",
-			"showName": true,
-			"tilesetId": null,
-			"tileId": null,
-			"tileRenderMode": "FitInside",
-			"maxCount": 1,
-			"limitScope": "PerLevel",
-			"limitBehavior": "PreventAdding",
-			"pivotX": 0,
-			"pivotY": 0,
-			"fieldDefs": []
-		}
-	], "tilesets": [
-		{
-			"__cWid": 8,
-			"__cHei": 8,
-			"identifier": "Tileset_0",
-			"uid": 2,
-			"relPath": "tileset_0.png",
-			"pxWid": 512,
-			"pxHei": 512,
-			"tileGridSize": 64,
-			"spacing": 0,
-			"padding": 0,
-			"tagsSourceEnumUid": null,
-			"enumTags": [],
-			"customData": [],
-			"savedSelections": [],
-			"cachedPixelData": {
-				"opaqueTiles": "1111111111111111111111111111111111111111111111111111111111111111",
-				"averageColors": "ffffffeeff99ff99ffbbffeeffffffffffffffffffffffeeff88ffaaffffffffffffffffffeeff99ffeeffffffffffffffffffffffddffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"
-			}
-		}
-	], "enums": [], "externalEnums": [], "levelFields": [] },
-	"levels": [
-		{
-			"identifier": "Level_0",
-			"uid": 0,
-			"worldX": 0,
-			"worldY": 0,
-			"pxWid": 256,
-			"pxHei": 256,
-			"__bgColor": "#696A79",
-			"bgColor": null,
-			"useAutoIdentifier": true,
-			"bgRelPath": null,
-			"bgPos": null,
-			"bgPivotX": 0.5,
-			"bgPivotY": 0.5,
-			"__bgPos": null,
-			"externalRelPath": null,
-			"fieldInstances": [],
-			"layerInstances": [
-				{
-					"__identifier": "Entities",
-					"__type": "Entities",
-					"__cWid": 16,
-					"__cHei": 16,
-					"__gridSize": 16,
-					"__opacity": 1,
-					"__pxTotalOffsetX": 0,
-					"__pxTotalOffsetY": 0,
-					"__tilesetDefUid": null,
-					"__tilesetRelPath": null,
-					"levelId": 0,
-					"layerDefUid": 3,
-					"pxOffsetX": 0,
-					"pxOffsetY": 0,
-					"visible": true,
-					"optionalRules": [],
-					"intGrid": [],
-					"intGridCsv": [],
-					"autoLayerTiles": [],
-					"seed": 464814,
-					"overrideTilesetUid": null,
-					"gridTiles": [],
-					"entityInstances": [
-						{
-							"__identifier": "Player",
-							"__grid": [5,5],
-							"__pivot": [0,0],
-							"__tile": null,
-							"width": 64,
-							"height": 32,
-							"defUid": 4,
-							"px": [80,80],
-							"fieldInstances": []
-						}
-					]
-				},
-				{
-					"__identifier": "Tiles",
-					"__type": "Tiles",
-					"__cWid": 4,
-					"__cHei": 4,
-					"__gridSize": 64,
-					"__opacity": 1,
-					"__pxTotalOffsetX": 0,
-					"__pxTotalOffsetY": 0,
-					"__tilesetDefUid": 2,
-					"__tilesetRelPath": "tileset_0.png",
-					"levelId": 0,
-					"layerDefUid": 1,
-					"pxOffsetX": 0,
-					"pxOffsetY": 0,
-					"visible": true,
-					"optionalRules": [],
-					"intGrid": [],
-					"intGridCsv": [],
-					"autoLayerTiles": [],
-					"seed": 3140661,
-					"overrideTilesetUid": null,
-					"gridTiles": [
-						{ "px": [0,0], "src": [192,128], "f": 0, "t": 19, "d": [0] },
-						{ "px": [64,0], "src": [0,192], "f": 0, "t": 24, "d": [1] },
-						{ "px": [128,0], "src": [128,192], "f": 0, "t": 26, "d": [2] },
-						{ "px": [192,0], "src": [0,192], "f": 0, "t": 24, "d": [3] },
-						{ "px": [0,64], "src": [0,192], "f": 0, "t": 24, "d": [4] },
-						{ "px": [64,64], "src": [128,192], "f": 0, "t": 26, "d": [5] },
-						{ "px": [128,64], "src": [128,192], "f": 0, "t": 26, "d": [6] },
-						{ "px": [192,64], "src": [128,192], "f": 0, "t": 26, "d": [7] },
-						{ "px": [0,128], "src": [128,192], "f": 0, "t": 26, "d": [8] },
-						{ "px": [64,128], "src": [0,192], "f": 0, "t": 24, "d": [9] },
-						{ "px": [128,128], "src": [0,192], "f": 0, "t": 24, "d": [10] },
-						{ "px": [192,128], "src": [0,192], "f": 0, "t": 24, "d": [11] },
-						{ "px": [0,192], "src": [0,192], "f": 0, "t": 24, "d": [12] },
-						{ "px": [64,192], "src": [128,64], "f": 0, "t": 10, "d": [13] },
-						{ "px": [128,192], "src": [128,192], "f": 0, "t": 26, "d": [14] },
-						{ "px": [192,192], "src": [0,192], "f": 0, "t": 24, "d": [15] }
-					],
-					"entityInstances": []
-				}
-			],
-			"__neighbours": []
-		}
-	]
-}
